fix type of data to satisfy the compiler

+++ src/persistentNodeCache.ts
@@ -223,7 +223,7 @@ export default class PersistentNodeCache extends NodeCache {
         const mode = 0o666;
 
         if(this.appendFileDescriptor) {
-            fs.write(this.appendFileDescriptor, data, 0, data.length, null, (writeErr, written, buffer) => {
+            fs.write(this.appendFileDescriptor, data as Uint8Array, 0, data.length, null, (writeErr, written, buffer) => {
                 if (writeErr) {
                     console.error('Error writing to file:', writeErr);
                 }
@@ -236,7 +236,7 @@ export default class PersistentNodeCache extends NodeCache {
                 return;
             }
             this.appendFileDescriptor = fd;
-            fs.write(fd, data, 0, data.length, null, (writeErr, written, buffer) => {
+            fs.write(fd, data as Uint8Array, 0, data.length, null, (writeErr, written, buffer) => {
                 if (writeErr) {
                     console.error('Error writing to file:', writeErr);
                 }
